+++ code/server/sv_cvars.c
@@ -117,19 +117,19 @@ SV_InitCvars(void)
   sv_hostname = Cvar_GetAndDescribe("sv_hostname", "noname", CVAR_SERVERINFO | CVAR_ARCHIVE, "Sets the name of the server.");
   sv_maxclients = Cvar_GetAndDescribe("sv_maxclients", "8", CVAR_SERVERINFO | CVAR_LATCH, "Maximum number of people allowed to join the server.");
   Cvar_CheckRange(sv_maxclients, "1", XSTRING(MAX_CLIENTS), CV_INTEGER);
-  sv_maxclientsPerIP = Cvar_GetAndDescribe("sv_maxclientsPerIP", "3", CVAR_ARCHIVE | CVAR_SERVERINFO | CVAR_LATCH, "Limits the number of simultaneous connections from the same IP address.");
+  sv_maxclientsPerIP = Cvar_GetAndDescribe("sv_maxclientsPerIP", "3", CVAR_ARCHIVE, "Limits the number of simultaneous connections from the same IP address.");
   Cvar_CheckRange(sv_maxclientsPerIP, "1", NULL, CV_INTEGER);
   sv_clientTLD = Cvar_GetAndDescribe("sv_clientTLD", "0", CVAR_ARCHIVE_ND, "Client country detection code.");
   Cvar_CheckRange(sv_clientTLD, NULL, NULL, CV_INTEGER);
-  sv_democlients = Cvar_GetAndDescribe("sv_democlients", "0", CVAR_SERVERINFO | CVAR_LATCH | CVAR_ARCHIVE, "Maximum number of people allowed to view serverside demos.");
+  sv_democlients = Cvar_GetAndDescribe("sv_democlients", "0", CVAR_SERVERINFO | CVAR_LATCH, "Maximum number of people allowed to view serverside demos.");
   sv_minRate = Cvar_GetAndDescribe("sv_minRate", "0", CVAR_ARCHIVE_ND | CVAR_SERVERINFO, "Minimum server bandwidth (in bit per second) a client can use.");
   Cvar_CheckRange(sv_minRate, "0", "100000", CV_INTEGER);
   sv_maxRate = Cvar_GetAndDescribe("sv_maxRate", "0", CVAR_ARCHIVE_ND | CVAR_SERVERINFO, "Maximum server bandwidth (in bit per second) a client can use.");
   Cvar_CheckRange(sv_maxRate, "0", "100000", CV_INTEGER);
   sv_dlRate = Cvar_GetAndDescribe("sv_dlRate", "1000", CVAR_ARCHIVE | CVAR_SERVERINFO, "Bandwidth allotted to PK3 file downloads via UDP, in kbyte/s.");
-  sv_maxOOBRate = Cvar_GetAndDescribe("sv_maxOOBRate", "20", CVAR_ARCHIVE | CVAR_SERVERINFO, "Max out of band rate for handling incoming packets.\nNOTE: sv_protect & SVP_XREAL must be set!");
+  sv_maxOOBRate = Cvar_GetAndDescribe("sv_maxOOBRate", "20", CVAR_ARCHIVE, "Max out of band rate for handling incoming packets.\nNOTE: sv_protect & SVP_XREAL must be set!");
   Cvar_CheckRange(sv_maxOOBRate, "1", "1000", CV_INTEGER);
-  sv_maxOOBRateIP = Cvar_GetAndDescribe("sv_maxOOBRateIP", "1", CVAR_ARCHIVE | CVAR_SERVERINFO, "Max out of band rate for handling incoming packets per IP address.\nNOTE: sv_protect & SVP_XREAL must be set!");
+  sv_maxOOBRateIP = Cvar_GetAndDescribe("sv_maxOOBRateIP", "1", CVAR_ARCHIVE, "Max out of band rate for handling incoming packets per IP address.\nNOTE: sv_protect & SVP_XREAL must be set!");
   Cvar_CheckRange(sv_maxOOBRateIP, "1", "1000", CV_INTEGER);
 #if defined(UDP_DOWNLOAD_OPTIMIZE)
   Cvar_CheckRange(sv_dlRate, "0", "1500", CV_INTEGER);
@@ -149,11 +149,11 @@ SV_InitCvars(void)
   Cvar_CheckRange(sv_warningscpu, "0", "100", CV_INTEGER);
   sv_warningsframetime = Cvar_GetAndDescribe("sv_warningsframetime", "30", CVAR_ARCHIVE, va("Sets the desired value the average frame time has to go over before warnings begin appearing.\nNOTE: Falls back to %i if not set!", FRAME_TIME_WARNING));
   Cvar_CheckRange(sv_warningsframetime, "0", "100", CV_INTEGER);
-  sv_floodWait = Cvar_GetAndDescribe("sv_floodWait", "500", CVAR_ARCHIVE | CVAR_SERVERINFO, "Time in milliseconds that a client has to wait before sending another client command.");
-  sv_floodLimit = Cvar_GetAndDescribe("sv_floodLimit", "8", CVAR_ARCHIVE | CVAR_SERVERINFO, "The number of client commands a client is allowed to send before flood protection triggers.");
+  sv_floodWait = Cvar_GetAndDescribe("sv_floodWait", "500", CVAR_ARCHIVE, "Time in milliseconds that a client has to wait before sending another client command.");
+  sv_floodLimit = Cvar_GetAndDescribe("sv_floodLimit", "8", CVAR_ARCHIVE, "The number of client commands a client is allowed to send before flood protection triggers.");
   sv_floodProtect = Cvar_GetAndDescribe("sv_floodProtect", "1", CVAR_ARCHIVE | CVAR_SERVERINFO, "Toggle server flood protection to keep players from bringing the server down.");
 #if defined(INCLUDE_SV_PINGFIX)
-  sv_pingFix = Cvar_GetAndDescribe("sv_pingFix", "1", CVAR_ARCHIVE | CVAR_SERVERINFO, "Fix client ping calculation to more accurately reflect packet loss and force minimum ping for humans to 1");
+  sv_pingFix = Cvar_GetAndDescribe("sv_pingFix", "1", CVAR_ARCHIVE, "Fix client ping calculation to more accurately reflect packet loss and force minimum ping for humans to 1");
 #endif
   sv_userInfoFloodProtect = Cvar_GetAndDescribe("sv_userInfoFloodProtect", "1", CVAR_ARCHIVE | CVAR_SERVERINFO, "Prevents users from flooding the server with userinfo changes by delaying the next change for 5 seconds.");
   sv_showAverageBPS = Cvar_GetAndDescribe("sv_showAverageBPS", "0", 0, "BSP Network debugging");
@@ -216,9 +216,9 @@ SV_InitCvars(void)
   sv_mapChecksum = Cvar_GetAndDescribe("sv_mapChecksum", "", CVAR_ROM, "Allows check for client server map to match.");
   sv_lanForceRate = Cvar_GetAndDescribe("sv_lanForceRate", "1", CVAR_ARCHIVE_ND, "Forces LAN clients to the maximum rate instead of accepting client setting.");
   sv_dequeuePeriod = Cvar_Get("sv_dequeuePeriod", "500", CVAR_ARCHIVE);
-  sv_protect = Cvar_GetAndDescribe("sv_protect", "7", CVAR_ARCHIVE | CVAR_SERVERINFO, "Sets the desired networking protection level and whether or not to print logs to the console.\n1 is equal to 0001 in binary, enabling SVP_XREAL.\n2 is equal to 0010 in binary, enabling SVP_OWOLF.\n4 is equal to 0100 in binary, enabling SVP_CONSOLE (console print)\nIf this is set to 3, 0011 in binary, it enables both SVP_XREAL and SVP_OWOLF, and does not print to console.\nIf this is set to 5, 0101 in binary, it enables SVP_XREAL and SVP_CONSOLE.\nIf this is set to 6, 0110 in binary, it enables SVP_OWOLF and SVP_CONSOLE.\nIf this is set to 7, 0111 in binary, it enables SVP_XREAL, SVP_OWOLF, and SVP_CONSOLE for all functionality.");
-  sv_protectLog = Cvar_GetAndDescribe("sv_protectLog", "sv_protect.log", CVAR_ARCHIVE | CVAR_SERVERINFO, "Sets the desired name of the sv_protect log file. To disable for developer print output, set to \"\".");
-  sv_protectLogInterval = Cvar_GetAndDescribe("sv_protectLogInterval", "1000", CVAR_ARCHIVE | CVAR_SERVERINFO, "Sets the desired time in milliseconds until the next write to sv_protectLog is allowed to happen.");
+  sv_protect = Cvar_GetAndDescribe("sv_protect", "7", CVAR_ARCHIVE, "Sets the desired networking protection level and whether or not to print logs to the console.\n1 is equal to 0001 in binary, enabling SVP_XREAL.\n2 is equal to 0010 in binary, enabling SVP_OWOLF.\n4 is equal to 0100 in binary, enabling SVP_CONSOLE (console print)\nIf this is set to 3, 0011 in binary, it enables both SVP_XREAL and SVP_OWOLF, and does not print to console.\nIf this is set to 5, 0101 in binary, it enables SVP_XREAL and SVP_CONSOLE.\nIf this is set to 6, 0110 in binary, it enables SVP_OWOLF and SVP_CONSOLE.\nIf this is set to 7, 0111 in binary, it enables SVP_XREAL, SVP_OWOLF, and SVP_CONSOLE for all functionality.");
+  sv_protectLog = Cvar_GetAndDescribe("sv_protectLog", "sv_protect.log", CVAR_ARCHIVE, "Sets the desired name of the sv_protect log file. To disable for developer print output, set to \"\".");
+  sv_protectLogInterval = Cvar_GetAndDescribe("sv_protectLogInterval", "1000", CVAR_ARCHIVE, "Sets the desired time in milliseconds until the next write to sv_protectLog is allowed to happen.");
   sv_owolfAffectsLan = Cvar_GetAndDescribe("sv_owolfAffectsLan", "0", CVAR_ARCHIVE, "Toggle whether or not sv_protect & SVP_OWOLF applies to lan clients.");
   sv_demoState = Cvar_Get("sv_demoState", "0", CVAR_ROM);
   sv_autoDemo = Cvar_Get("sv_autoDemo", "0", CVAR_ARCHIVE);